2.0

+++ main.tf
@@ -95,12 +95,14 @@ resource "null_resource" "REMOVE_KEY" {
 
     for_each = local.KEYs
     triggers = {
-        KEY_PRI_WIN_FILE = each.value.KEY_PRI_WIN_FILE
-        KEY_PRI_LINUX_FILE = each.value.KEY_PRI_LINUX_FILE
-        KEY_PRI_S3_FILE =  each.value.KEY_PRI_S3_FILE
-        KEY_PUB_WIN_FILE = each.value.KEY_PUB_WIN_FILE
-        KEY_PUB_LINUX_FILE = each.value.KEY_PUB_LINUX_FILE
-        KEY_PUB_S3_FILE =  each.value.KEY_PUB_S3_FILE
+        KEY_PRI_WIN_FILE = try(each.value.KEY_PRI_WIN_FILE, "")
+        KEY_PRI_LINUX_FILE = try(each.value.KEY_PRI_LINUX_FILE, "")
+        KEY_PRI_RUNNER_FILE =  try(each.value.KEY_PRI_RUNNER_FILE, "")
+        KEY_PRI_S3_FILE =  try(each.value.KEY_PRI_S3_FILE, "")
+        KEY_PUB_WIN_FILE = try(each.value.KEY_PUB_WIN_FILE, "")
+        KEY_PUB_LINUX_FILE = try(each.value.KEY_PUB_LINUX_FILE, "")
+        KEY_PUB_RUNNER_FILE =  try(each.value.KEY_PUB_RUNNER_FILE, "")
+        KEY_PUB_S3_FILE =  try(each.value.KEY_PUB_S3_FILE, "")
         PROFILE = "${var.PROFILE}"
     }
 
@@ -110,27 +112,39 @@ resource "null_resource" "REMOVE_KEY" {
         command = <<-EOF
             if [ -f "${self.triggers.KEY_PRI_WIN_FILE}" ]; then
                 sudo rm -rf "${self.triggers.KEY_PRI_WIN_FILE}"
-                aws s3 rm "s3://${self.triggers.KEY_PRI_S3_FILE}" --profile ${self.triggers.PROFILE})
+                if [ -f "${self.triggers.KEY_PRI_S3_FILE}" ]; then
+                    aws s3 rm "s3://${self.triggers.KEY_PRI_S3_FILE}" --profile ${self.triggers.PROFILE})
+                fi
             fi
             if [ -f "${self.triggers.KEY_PUB_WIN_FILE}" ]; then
                 sudo rm -rf "${self.triggers.KEY_PUB_WIN_FILE}"
-                aws s3 rm "s3://${self.triggers.KEY_PUB_S3_FILE}" --profile ${self.triggers.PROFILE})
+                if [ -f "${self.triggers.KEY_PUB_S3_FILE}" ]; then
+                    aws s3 rm "s3://${self.triggers.KEY_PUB_S3_FILE}" --profile ${self.triggers.PROFILE})
+                fi
             fi
             if [ -f "${self.triggers.KEY_PRI_LINUX_FILE}" ]; then
                 sudo rm -rf "${self.triggers.KEY_PRI_LINUX_FILE}"
-                aws s3 rm "s3://${self.triggers.KEY_PRI_S3_FILE}" --profile ${self.triggers.PROFILE})
+                if [ -f "${self.triggers.KEY_PRI_S3_FILE}" ]; then
+                    aws s3 rm "s3://${self.triggers.KEY_PRI_S3_FILE}" --profile ${self.triggers.PROFILE})
+                fi
             fi
             if [ -f "${self.triggers.KEY_PUB_LINUX_FILE}" ]; then
                 sudo rm -rf "${self.triggers.KEY_PUB_LINUX_FILE}"
-                aws s3 rm "s3://${self.triggers.KEY_PUB_S3_FILE}" --profile ${self.triggers.PROFILE})
+                if [ -f "${self.triggers.KEY_PUB_S3_FILE}" ]; then
+                    aws s3 rm "s3://${self.triggers.KEY_PUB_S3_FILE}" --profile ${self.triggers.PROFILE})
+                fi
             fi
             if [ -f "${self.triggers.KEY_PRI_RUNNER_FILE}" ]; then
                 sudo rm -rf "${self.triggers.KEY_PRI_RUNNER_FILE}"
-                aws s3 rm "s3://${self.triggers.KEY_PRI_S3_FILE}" --profile ${self.triggers.PROFILE})
+                if [ -f "${self.triggers.KEY_PRI_S3_FILE}" ]; then
+                    aws s3 rm "s3://${self.triggers.KEY_PRI_S3_FILE}" --profile ${self.triggers.PROFILE})
+                fi
             fi
             if [ -f "${self.triggers.KEY_PUB_RUNNER_FILE}" ]; then
                 sudo rm -rf "${self.triggers.KEY_PUB_RUNNER_FILE}"
-                aws s3 rm "s3://${self.triggers.KEY_PUB_S3_FILE}" --profile ${self.triggers.PROFILE})
+                if [ -f "${self.triggers.KEY_PUB_S3_FILE}" ]; then
+                    aws s3 rm "s3://${self.triggers.KEY_PUB_S3_FILE}" --profile ${self.triggers.PROFILE})
+                fi
             fi
         EOF
     }
